Fix SqlTrack playcount semantics towards LastFM/FMPS rules.
Max, you said that this would have to be implemented in at least two places, but unless you mean in StatisticsProvider classes (which don't know the length of a track) I didn't find another place.

CCMAIL:amarok-devel@kde.org

+++ src/collection/sqlcollection/SqlMeta.h
@@ -183,7 +183,7 @@ class SqlTrack : public Meta::Track
         QString m_rpath;
         int m_trackId;
 
-        int m_length;
+        qint64 m_length;
         qint64 m_filesize;
         int m_trackNumber;
         int m_discNumber;
